metasfresh 5.168

Release notes can be found at https://github.com/metasfresh/metasfresh/blob/master/ReleaseNotes.md#metasfresh-5168

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/picking/pickingslot/process/WEBUI_Picking_ReturnQtyToSourceHU.java
@@ -1,24 +1,42 @@
 package de.metas.ui.web.picking.pickingslot.process;
 
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_CU;
-import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_HU;
-
-import java.math.BigDecimal;
-import java.util.Collection;
-import java.util.Objects;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
+import com.google.common.collect.ImmutableList;
 import de.metas.handlingunits.HuId;
+import de.metas.handlingunits.IHandlingUnitsBL;
+import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.picking.PickingCandidateService;
 import de.metas.handlingunits.picking.requests.RemoveQtyFromHURequest;
 import de.metas.handlingunits.sourcehu.HuId2SourceHUsService;
+import de.metas.handlingunits.storage.IHUStorage;
+import de.metas.handlingunits.storage.IHUStorageFactory;
+import de.metas.handlingunits.storage.IProductStorage;
+import de.metas.process.IProcessDefaultParameter;
+import de.metas.process.IProcessDefaultParametersProvider;
 import de.metas.process.IProcessPrecondition;
 import de.metas.process.Param;
 import de.metas.process.ProcessPreconditionsResolution;
+import de.metas.product.ProductId;
 import de.metas.ui.web.handlingunits.HUEditorRowType;
 import de.metas.ui.web.picking.pickingslot.PickingSlotRow;
+import de.metas.ui.web.process.descriptor.ProcessParamLookupValuesProvider;
+import de.metas.ui.web.window.datatypes.LookupValuesList;
+import de.metas.ui.web.window.descriptor.DocumentLayoutElementFieldDescriptor;
+import de.metas.ui.web.window.model.lookup.LookupDataSourceFactory;
+import de.metas.util.Services;
+import lombok.NonNull;
+import org.adempiere.exceptions.FillMandatoryException;
+import org.compiere.SpringContextHolder;
+import org.compiere.model.I_M_Product;
+
+import javax.annotation.Nullable;
+import java.math.BigDecimal;
+import java.util.Collection;
+import java.util.List;
+import java.util.Objects;
+
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_MISSING_SOURCE_HU;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_CU;
+import static de.metas.ui.web.picking.PickingConstants.MSG_WEBUI_PICKING_SELECT_PICKED_HU;
 
 /*
  * #%L
@@ -44,17 +62,19 @@ import de.metas.ui.web.picking.pickingslot.PickingSlotRow;
 
 public class WEBUI_Picking_ReturnQtyToSourceHU
 		extends WEBUI_Picking_With_M_Source_HU_Base
-		implements IProcessPrecondition
+		implements IProcessPrecondition, IProcessDefaultParametersProvider
 {
-	@Autowired
-	private PickingCandidateService pickingCandidateService;
-
-	@Autowired
-	private HuId2SourceHUsService sourceHUsRepository;
+	private final PickingCandidateService pickingCandidateService = SpringContextHolder.instance.getBean(PickingCandidateService.class);
+	private final HuId2SourceHUsService sourceHUsRepository = SpringContextHolder.instance.getBean(HuId2SourceHUsService.class);
+	private final IHandlingUnitsBL handlingUnitsBL = Services.get(IHandlingUnitsBL.class);
 
 	private static final String PARAM_QTY_CU = "QtyCU";
 	@Param(parameterName = PARAM_QTY_CU, mandatory = true)
 	private BigDecimal qtyCU;
+
+	private static final String PARAM_M_Product_ID = "M_Product_ID";
+	@Param(parameterName = PARAM_M_Product_ID, mandatory = true)
+	private ProductId productId;
 
 	@Override
 	protected ProcessPreconditionsResolution checkPreconditionsApplicable()
@@ -85,15 +105,51 @@ public class WEBUI_Picking_ReturnQtyToSourceHU
 		return ProcessPreconditionsResolution.accept();
 	}
 
+	@Nullable
 	@Override
-	protected String doIt() throws Exception
+	public Object getParameterDefaultValue(final IProcessDefaultParameter parameter)
 	{
-		final PickingSlotRow pickingSlotRow = getSingleSelectedRow();
+		if (PARAM_M_Product_ID.equals(parameter.getColumnName()))
+		{
+			final List<ProductId> productIds = getProductIds(getSelectedHUId());
+			return !productIds.isEmpty() ? productIds.get(0) : null;
+		}
+		else
+		{
+			return IProcessDefaultParametersProvider.DEFAULT_VALUE_NOTAVAILABLE;
+		}
+	}
+
+	@ProcessParamLookupValuesProvider(
+			parameterName = PARAM_M_Product_ID,
+			numericKey = true,
+			lookupSource = DocumentLayoutElementFieldDescriptor.LookupSource.list,
+			lookupTableName = I_M_Product.Table_Name)
+	@NonNull
+	public LookupValuesList getAvailableProducts()
+	{
+		final List<ProductId> productIds = getProductIds(getSelectedHUId());
+		return LookupDataSourceFactory.instance.searchInTableLookup(I_M_Product.Table_Name).findByIdsOrdered(productIds);
+	}
+
+	@Override
+	protected String doIt()
+	{
+		if (productId == null)
+		{
+			throw new FillMandatoryException(PARAM_M_Product_ID);
+		}
+		if (qtyCU == null || qtyCU.signum() <= 0)
+		{
+			throw new FillMandatoryException(PARAM_QTY_CU);
+		}
+
+		final HuId huId = getSelectedHUId();
 
 		pickingCandidateService.removeQtyFromHU(RemoveQtyFromHURequest.builder()
 				.qtyCU(qtyCU)
-				.huId(pickingSlotRow.getHuId())
-				.productId(pickingSlotRow.getHuProductId())
+				.huId(huId)
+				.productId(productId)
 				.build());
 
 		invalidateView();
@@ -102,10 +158,29 @@ public class WEBUI_Picking_ReturnQtyToSourceHU
 		return MSG_OK;
 	}
 
+	private HuId getSelectedHUId()
+	{
+		return getSingleSelectedRow().getHuId();
+	}
+
 	private boolean checkSourceHuPreconditionIncludingEmptyHUs()
 	{
-		final HuId huId = getSingleSelectedRow().getHuId();
+		final HuId huId = getSelectedHUId();
 		final Collection<HuId> sourceHUs = sourceHUsRepository.retrieveMatchingSourceHUIds(huId);
 		return !sourceHUs.isEmpty();
+	}
+
+	private List<ProductId> getProductIds(@NonNull final HuId huId)
+	{
+		final I_M_HU hu = handlingUnitsBL.getById(huId);
+		final IHUStorageFactory storageFactory = handlingUnitsBL.getStorageFactory();
+		final IHUStorage storage = storageFactory.getStorage(hu);
+
+		return storage.getProductStorages()
+				.stream()
+				.filter(productStorage -> !productStorage.isEmpty())
+				.map(IProductStorage::getProductId)
+				.distinct()
+				.collect(ImmutableList.toImmutableList());
 	}
 }
